==> 5-Text-Generation.ipynb <==
# Reading the corpus
import pandas as pd

data = pd.read_pickle('corpus.pkl')
data

# Ali Wong's text
ali_text = data.transcript.loc['ali']
ali_text[:200]

# Building a Markov Chain Function

from collections import defaultdict

def markov_chain(text):
    '''The input is a string of text and the output will be a dictionary with each word as
       a key and each value as the list of words that come after the key in the text.'''
    
    # Tokenizing the text by word, including punctuation
    words = text.split(' ')
    
    # Initializing a default dictionary to hold all of the words and next words
    m_dict = defaultdict(list)
    
    # Creating a zipped list of all of the word pairs and put them in word: list of next words format
    for current_word, next_word in zip(words[0:-1], words[1:]):
        m_dict[current_word].append(next_word)

    # Converting the default dict back into a dictionary
    m_dict = dict(m_dict)
    return m_dict

# Creating the dictionary for Ali's routine
ali_dict = markov_chain(ali_text)
ali_dict

# Create a Text Generator

import random

def generate_sentence(chain, count=15):
    '''Input a dictionary in the format of key = current word, value = list of next words
       along with the number of words you would like to see in your generated sentence.'''

    # Capitalizing the first word
    word1 = random.choice(list(chain.keys()))
    sentence = word1.capitalize()

    # Generating the second word from the value list
    for i in range(count-1):
        word2 = random.choice(chain[word1])
        word1 = word2
        sentence += ' ' + word2

    # Ending it with a period
    sentence += '.'
    return(sentence)

generate_sentence(ali_dict)
